Prefer mode of type "DRM_MODE_TYPE_PREFERED"
On screen with native resolution of 1440x900, there is sometime a
compatibility mode for 1280x1024, with shrinked output and vertical
bands to maintain ratio. If we select the resolution with the
higher area, we select the compatibility mode and not the native
resolution.

When using mode_type == DRM_MODE_TYPE_PREFERED, we avoid that issue.

Signed-off-by: Gabriel Laskar <gabriel@lse.epita.fr>