Make search for a crtc more robust
Running the kmscube test on qemu with "-vga virtio" fails for me with
a "no encoder!" error.

Fixed by searching all crtcs for one that is supported by one of the
encoders in the connector.

This code is only used if the existing `encoder->crtc_id` check is
zero, so there should be no regression for systems where it already
works. I think this should match the advice given here:

https://dvdhrm.wordpress.com/2012/09/13/linux-drm-mode-setting-api/