Fix GBM image formats v2
When the GBM BOs are created, they are created using a bo with a correct
memory requirement but with an incorrect channel layout. This makes
drivers that are picky about matching formats (svga/vmwgfx) complain.

Use the correct GBM formats and require GBM >= 13.0 to make sure they
are available.

v2: Remove a couple of stale comments

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Emil Velikov <emil.velikov@collabora.com>