atomic: Use CPU-side KMS fence synchronisation
Atomic does not let us have multiple requests in flight at once; if we
don't synchronise with the request completion on the CPU side, we can
get -EBUSY from the atomic commit.

We already have everything required to do this, namely the out-fence
from the last commit. Block on that with a CPU-side wait before we call
atomic commit, to make sure we'll never attempt to queue an atomic
commit before the previous one has completed.

Signed-off-by: Daniel Stone <daniels@collabora.com>
Reviewed-by: Emil Velikov <emil.velikov@collabora.com>