WIP: video: fencing to avoid passing frame back to decoder too early
With atomic kms backend, userspace can get further ahead of the gpu.  In
the decoder, we unref the previous frame when retrieving the current
frame.  If userspace gets too far ahead, this can happen while the gpu
is still sampling from the previous frame.

Simple solution is add a fence.