common: Re-support EGL 1.4 by using EGLImageKHR when needed
EGLImage was added to core EGL in version 1.5, but was available as an
extension in 1.4 as EGLImageKHR, all of the rest of kmscube is EGL 1.4
compatible except this typedef. Allow kmscube to be EGL 1.4 compatible
again by checking for EGL 1.5 and if not available falling back to the
old typedef name.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Reviewed-by: Eric Engestrom <eric.engestrom@intel.com>
Reviewed-by: Emil Velikov <emil.velikov@collabora.com>